pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "autogen>=0.9.3",
    "autogen-agentchat>=0.6.1",
    "autogen-ext>=0.6.1",
    "mcp[cli]>=1.6.0",
    "openai>=1.91.0",
    "pandas>=2.2.3",
    "pyarrow>=19.0.1",
    "python-dotenv>=1.1.1",
]
